refactor: reorganize imports and enhance Header component styling
- Updated the Header component to improve the organization of imports, ensuring that hooks and components are clearly separated for better readability.
- Adjusted the CSS classes in the Header to maintain consistent styling, enhancing the visual appearance and user experience.
- Refactored the use of the `transformProductData` function by moving it to a new `productTransformer.js` file, improving the modularity of the codebase.
- Removed the obsolete `productService.js` file, streamlining the services used in the application.
- Enhanced the AllProduct and SearchResults pages by ensuring consistent import statements and improving the overall structure of the code.
- This change aims to improve code maintainability and readability while ensuring a consistent styling approach across components.
